pyproject.toml
[project]
name = "GONG-ACCOUNT-SUMMARY"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.6.0",
    "anthropic>=0.34.0",
    "google-api-python-client>=2.147.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]
